# src/PaPl.SKS.Package.Services/wwwroot/swagger-original.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Parcel Logistics Service",
    "contact": {
      "name": "SKS",
      "url": "http://www.technikum-wien.at/"
    },
    "version": "1.20.1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/parcel/{trackingId}/reportHop/{code}": {
      "post": {
        "operationId": "reportParcelHop",
        "summary": "Report that a Parcel has arrived at a certain hop either Warehouse or Truck.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/trackingId"
          },
          {
            "$ref": "#/components/parameters/code"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully reported hop."
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "description": "Parcel does not exist with this tracking ID or hop with code not found.\n"
          }
        },
        "tags": [
          "staff"
        ]
      }
    },
...
  },
  "components": {
    "parameters": {
      "code": {
        "name": "code",
        "in": "path",
        "required": true,
        "schema": {
          "pattern": "^[A-Z0-9]{9}$",
          "type": "string"
        },
        "description": "The Code of the hop (Warehouse or Truck)."
      },
      "trackingId": {
        "name": "trackingId",
        "in": "path",
        "description": "The tracking ID of the parcel.\nE.g. PYJRB4HZ6\n",
        "required": true,
        "schema": {
          "pattern": "^[A-Z0-9]{9}$",
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "error": {
        "required": [
          "errorMessage"
        ],
        "type": "object",
        "properties": {
          "errorMessage": {
            "type": "string",
            "description": "The error message."
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "The operation failed due to an error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        }
      }
    }
  }
}